test: 리팩토링에 따른 수익률 계산 테스트 수정
